+++ main.py
@@ -1,4 +1,4 @@
-from src.main import poll_and_notify
+from src.web_differ import poll_and_notify
 
 if __name__ == "__main__":
     poll_and_notify()
